% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcc-stats.R
\name{tcc_stats}
\alias{tcc_stats}
\title{Stats}
\usage{
tcc_stats(parse = TRUE)
}
\arguments{
\item{parse}{Logical indicating whether to convert the retrieved JSON data
into a tidy tibble data frame. Default is TRUE.}
}
\value{
A list of four tibbles (data frames) with various meta data and
Twitter-related data about members of the current U.S. Congress.
}
\description{
Returns stats data from TweetCongress 'Tweets' API
}
\examples{

\dontrun{
## get tweeters data
stats <- tcc_stats()
}

}
