% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcc-tweeters.R
\name{tcc_tweeters}
\alias{tcc_tweeters}
\title{Tweeters}
\usage{
tcc_tweeters(parse = TRUE)
}
\arguments{
\item{parse}{Logical indicating whether to convert the retrieved JSON data
into a tidy tibble data frame. Default is TRUE.}
}
\value{
A tibble (data frame) with various meta data and Twitter-related data
about members of the current U.S. Congress.
}
\description{
Returns tweeter data from TweetCongress 'Tweeters' API
}
\examples{

\dontrun{
## get tweeters data
tweeters <- tcc_tweeters()
}

}
