% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcc-tweets.R
\name{tcc_tweets}
\alias{tcc_tweets}
\title{Tweets}
\usage{
tcc_tweets(parse = TRUE)
}
\arguments{
\item{parse}{Logical indicating whether to convert the retrieved JSON data
into a tidy tibble data frame. Default is TRUE.}
}
\value{
A tibble (data frame) with various meta data and Twitter-related data
about members of the current U.S. Congress.
}
\description{
Returns tweets data from TweetCongress 'Tweets' API
}
\examples{

\dontrun{
## get tweeters data
tweets <- tcc_tweets()
}

}
