% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twcong.R
\name{twcong_parse_data}
\alias{twcong_parse_data}
\title{Parse TweetCongress data}
\usage{
twcong_parse_data(x)
}
\arguments{
\item{x}{Input data–this should be the JSON data retrieved from TweetCongress}
}
\value{
A tibble (data frame) with information about Congress and Twitter.
}
\description{
Converts JSON data returned by TweetCongress into tidy tibble (data frames)
}
\examples{

\dontrun{
## get tweeters data
twt <- twcong_tweeters(parse = FALSE)

## see it's a big, recursive list
str(twt, 3)

## now it's a nice tibble
twcong_parse_data(twt)
}

}
