% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twcong.R
\name{twcong_tweeters}
\alias{twcong_tweeters}
\title{Tweeters}
\usage{
twcong_tweeters(parse = TRUE)
}
\arguments{
\item{parse}{Logical indicating whether to convert the retrieved JSON data
into a tidy tibble data frame. Default is TRUE.}
}
\value{
A tibble (data frame) with various meta data and Twitter-related data
about members of the current U.S. Congress.
}
\description{
Returns data from TweetCongress 'Tweeters' API
}
\examples{

\dontrun{
## get tweeters data
twt <- twcong_tweeters()
}

}
